requirements.txt
Flask==2.0.3
psycopg2==2.9.3
gunicorn==20.1.0
pandas<=1.3.5
beautifulsoup4==4.10.0
clickhouse-driver[numpy,lz4]==0.2.3
Flask-Cors
Flask-SQLAlchemy
requests
sentry-sdk[flask]
python-dateutil
protobuf==3.20.1
google-cloud-bigquery
google-cloud-logging>=3.0.0
pycodestyle
git+https://github.com/ClickHouse/clickhouse-connect.git@0452f496df15cc0410be94c5f6fa01ddf2c25c76#egg=clickhouse-connect
#clickhouse_connect
# polygon-etl
# google shit
google-cloud-secret-manager==2.10.0
solana